Split minipyro into contrib + example
