+++ docs/source/contrib.minipyro.rst
@@ -1,0 +1,5 @@
+.. automodule:: pyro.contrib.minipyro
+    :members:
+    :undoc-members:
+    :special-members: __call__
+    :show-inheritance:
