+++ tests/test_examples.py
@@ -36,7 +36,8 @@ CPU_EXAMPLES = [
     'hmm.py --num-steps=1 --truncate=65 --model=5',
     'inclined_plane.py --num-samples=1',
     'lda.py --num-steps=2 --num-words=100 --num-docs=100 --num-words-per-doc=8',
-    'mini_pyro.py',
+    'minipyro.py',
+    'minipyro.py --full',
     'rsa/generics.py --num-samples=10',
     'rsa/hyperbole.py --price=10000',
     'rsa/schelling.py --num-samples=10',
