+++ tutorial/source/mini_pyro.py
@@ -1,0 +1,1 @@
+../../examples/mini_pyro.py
